同步位置;

+++ client/Trunk/Assets/Scripts/MapModules/Charcter/Data/VO/WorldRoleInfoVO.cs
@@ -54,7 +54,7 @@ public class WorldRoleInfoVO
 
 public class ModelPosVO
 {
-    public int posX;
-    public int posY;
+    public double posX;
+    public double posY;
     public int angle;
 }
